refactor: 코드 리뷰 반영 - 로그인 타입 파일 만듦 - logout 중복 선언 정리 - myPage 탈퇴 추가, p 태그 -> 버튼 - useAuthState 상태 가져오는 코드 수정
