+++ package.json
@@ -19,7 +19,6 @@
     "@tanstack/react-query": "^5.66.0",
     "axios": "^1.7.9",
     "gsap": "^3.12.7",
-    "immer": "^10.1.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-intersection-observer": "^9.15.1",
