+++ src/types/auth.d.ts
@@ -1,0 +1,1 @@
+type LoginType = 'kakao' | 'naver' | 'google';
